requirements.txt
# === 기존 의존성 (web_app.py 용) ===
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2
html5lib==1.1
flask==3.0.0
markdown==3.5.1

# === 신규 의존성 (main.py - 문서 분석 웹앱) ===
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
PyPDF2==3.0.1
python-docx==1.1.2
openai==1.51.0
anthropic==0.34.2
python-dotenv==1.0.1
